fix: change struct dupcheck type data int to float

+++ models/response/response.go
@@ -90,7 +90,7 @@ type DataDupcheck struct {
 	NumOfAssetInventoried            int     `json:"num_of_asset_inventoried"`
 	NumOfDependence                  int     `json:"num_of_dependence"`
 	NumberOfAgreement                int     `json:"number_of_agreement"`
-	OsInstallmentdue                 int     `json:"os_installmentdue"`
+	OsInstallmentdue                 float64 `json:"os_installmentdue"`
 	OutstandingPrincipal             float64 `json:"outstanding_principal"`
 	OverduedaysAging                 int     `json:"overduedays_aging"`
 	PersonalNpwp                     string  `json:"personal_npwp"`
@@ -102,10 +102,10 @@ type DataDupcheck struct {
 	ResidenceZipcode                 string  `json:"residence_zipcode"`
 	RrdDate                          string  `json:"rrd_date"`
 	SisaJumlahAngsuran               int     `json:"sisa_jumlah_angsuran"`
-	SpouseIncome                     int     `json:"spouse_income"`
+	SpouseIncome                     float64 `json:"spouse_income"`
 	SurgateMotherName                string  `json:"surgate_mother_name"`
-	TotalInstallment                 int     `json:"total_installment"`
-	TotalInstallmentNap              int     `json:"total_installment_nap"`
+	TotalInstallment                 float64 `json:"total_installment"`
+	TotalInstallmentNap              float64 `json:"total_installment_nap"`
 	WorkSinceYear                    string  `json:"work_since_year"`
 	InstallmentAmount_ChassisNo      string  `json:"installment_amount_chassis_no"`
 }
